refactor: EntityRowMapper 추상화 변경
